add in-memory fallback for PvFile
